Cleaning up and decoupling
