+++ views/actions/Camera.js
@@ -1,6 +1,9 @@
 import ACTIONS from '../actions';
 import KEYS from '../keys';
 
+import { nextCameraPositionToPoint, translateToPointInCamera, nextPositionToPointToLookAt } from '../helpers/Movement.js';
+
+
 const mapCharToDirection = (key) => {
   switch (key) {
     case KEYS.I:
@@ -21,11 +24,30 @@ const mapCharToDirection = (key) => {
 export const cameraRotation = key => (dispatch, getState) => {
   const { Character } = getState();
   return dispatch({
-    type: ACTIONS.CAMERA.ROTATE,
+    type: ACTIONS.CAMERA.ROTATE_CAMERA,
     direction: mapCharToDirection(key),
     speed: 5,
     characterPosition: Character.characterPosition,
   });
+};
+
+export const cameraMoveToPointSuccess = cameraNextPosition => ({
+  type: ACTIONS.CAMERA.MOVE_CAMERA,
+  cameraNextPosition: cameraNextPosition,
+});
+
+export const cameraMoveToPoint = point => (dispatch, getState) => {
+  const { Character, Camera } = getState();
+
+  const characterPositionTranslated = translateToPointInCamera(Character.characterPosition);
+  const cameraNextPosition = {
+    lookAt: nextPositionToPointToLookAt(Camera.cameraConfig.lookAt, point, Character.config.speed),
+    position: nextCameraPositionToPoint(
+      Camera.cameraConfig.position, characterPositionTranslated, point, Character.config.speed
+    ),
+  };
+
+  dispatch(cameraMoveToPointSuccess(cameraNextPosition));
 };
 
 export default cameraRotation;
