+++ views/actions/Character.js
@@ -1,10 +1,10 @@
 import ACTIONS from '../actions';
-import { nextPositionToPoint, nextCameraPositionToPoint, translateToPointInMap, translateToPointInCamera, nextPositionToPointToLookAt } from '../helpers/Movement.js';
+import { nextPositionToPoint, translateToPointInMap } from '../helpers/Movement.js';
+import { cameraMoveToPoint } from './Camera';
 
-export const characterMoveToPointSuccess = (characterNextPosition, cameraNextPosition) => ({
+export const characterMoveToPointSuccess = characterNextPosition => ({
   type: ACTIONS.CHARACTER.WALK,
   characterNextPosition: characterNextPosition,
-  cameraNextPosition: cameraNextPosition,
 });
 
 export const characterMoveToPoint = point => (dispatch, getState) => {
@@ -12,26 +12,18 @@ export const characterMoveToPoint = point => (dispatch, getState) => {
     return;
   }
 
-  const { Character, Camera } = getState();
+  const { Character } = getState();
 
   const characterPreviousPosition = Character.characterPosition;
   const characterNextPosition = nextPositionToPoint(
     characterPreviousPosition, translateToPointInMap(point), Character.config.speed
   );
 
-  const cameraLookAtPosition = Camera.cameraConfig.lookAt;
-  const cameraPreviousPosition = Camera.cameraConfig.position;
-  const cameraNextPosition = {
-    lookAt: nextPositionToPointToLookAt(cameraLookAtPosition, point, Character.config.speed),
-    position: nextCameraPositionToPoint(
-      cameraPreviousPosition, translateToPointInCamera(characterNextPosition), point, Character.config.speed
-    ),
-  };
-
   const hasCharacterMoved = !Character.characterPosition.equals(characterNextPosition);
 
   if (hasCharacterMoved) {
-    dispatch(characterMoveToPointSuccess(characterNextPosition, cameraNextPosition));
+    dispatch(characterMoveToPointSuccess(characterNextPosition));
+    dispatch(cameraMoveToPoint(point));
   }
 };
